+++ SplatTagCore/Division.cs
@@ -1,77 +1,20 @@
-﻿namespace SplatTagCore
+﻿using System.Collections.Generic;
+
+namespace SplatTagCore
 {
   /// <summary>
-  /// A division. X is div 0. X+ is div -1. Unknown is int.MaxValue.
-  /// Higher divs are LOWER in number.
+  /// A division.
   /// </summary>
-  public class Division
+  public interface IDivision
   {
-    public const int X_PLUS = -1;
-    public const int X = 0;
-    public const int UNKNOWN = int.MaxValue;
+    /// <summary>
+    /// Get a string name or representation of this div.
+    /// </summary>
+    string Name { get; }
 
-    public static Division Unknown = new Division();
-    public int div = UNKNOWN;
-
-    public Division(int div = UNKNOWN)
-    {
-      this.div = div;
-    }
-
-    public Division(string div)
-    {
-      if (div == null)
-      {
-        this.div = UNKNOWN;
-      }
-      else
-      {
-        div = div.Trim(new char[] { '(', 'D', 'd', 'i', 'v', ')', ' ', });
-
-        if (string.IsNullOrWhiteSpace(div))
-        {
-          // Unknown
-          this.div = int.MaxValue;
-        }
-        else if (div.Equals("X+", System.StringComparison.OrdinalIgnoreCase))
-        {
-          this.div = X_PLUS;
-        }
-        else if (div.Equals("X", System.StringComparison.OrdinalIgnoreCase))
-        {
-          this.div = X;
-        }
-        else if (int.TryParse(div, out int divParse))
-        {
-          this.div = divParse;
-        }
-        else if (int.TryParse(div[0].ToString(), out int divParse2))
-        {
-          // 8U and 8M, hopefully will protect against mixed divisions in future.
-          // May cause a problem for strings such as 10M (if we get to 10+ divisions).
-          this.div = divParse2;
-        }
-        else
-        {
-          this.div = UNKNOWN;
-        }
-      }
-    }
-
-    public static implicit operator int(Division d)
-    {
-      return d.div;
-    }
-
-    public override string ToString()
-    {
-      switch (div)
-      {
-        case UNKNOWN: return "Unknown";
-        case X_PLUS: return "X+";
-        case X: return "X";
-        default: return div.ToString();
-      }
-    }
+    /// <summary>
+    /// Get the value of this div.
+    /// </summary>
+    int Value { get; }
   }
 }