+++ SplatTagDatabase/Importers/LUTIJsonReader.cs
@@ -20,8 +20,9 @@ namespace SplatTagDatabase.Importers
       [JsonProperty("Division", Required = Required.Default)]
       public string Division { get; set; } = "Unknown";
 
+      private string tag;
       [JsonProperty("Tag")]
-      public string Tag { get; set; }
+      public string Tag { get => tag; set => tag = value.Trim(); }
 
       [JsonProperty("Team Captain", Required = Required.Default)]
       public string TeamCaptain { get; set; } = "";
@@ -71,7 +72,7 @@ namespace SplatTagDatabase.Importers
         throw new InvalidOperationException("jsonFile is not set.");
       }
 
-      void CheckAndAddPlayer(string tryPlayerName, string _tag, Team _newTeam, List<Player> _players)
+      void CheckAndAddPlayer(string tryPlayerName, string _tag, string _transformedTag, Team _newTeam, List<Player> _players)
       {
         if (!string.IsNullOrWhiteSpace(tryPlayerName))
         {
@@ -82,12 +83,20 @@ namespace SplatTagDatabase.Importers
               {
                 tryPlayerName = tryPlayerName.Substring(_tag.Length).Trim();
               }
+              else if (tryPlayerName.StartsWith(_transformedTag))
+              {
+                tryPlayerName = tryPlayerName.Substring(_transformedTag.Length).Trim();
+              }
               break;
 
             case TagOption.Back:
               if (tryPlayerName.EndsWith(_tag))
               {
                 tryPlayerName = tryPlayerName.Substring(0, tryPlayerName.Length - _tag.Length - 1).Trim();
+              }
+              else if (tryPlayerName.EndsWith(_transformedTag))
+              {
+                tryPlayerName = tryPlayerName.Substring(0, tryPlayerName.Length - _transformedTag.Length - 1).Trim();
               }
               break;
 
@@ -122,50 +131,62 @@ namespace SplatTagDatabase.Importers
           Id = (uint)teams.Count,  // This will be updated when the merge happens.
           ClanTags = new string[] { row.Tag },
           ClanTagOption = TagOption.Front,
-          Div = new Division(row.Division),
+          Div = new LUTIDivision(row.Division),
           Name = row.TeamName,
           Sources = new List<string> { Path.GetFileNameWithoutExtension(jsonFile) }
         };
 
         // Handle tag placements from the captain's name
-        if (string.IsNullOrWhiteSpace(row.Tag))
+        string transformedTag = row.Tag?.TransformString();
+        if (string.IsNullOrWhiteSpace(transformedTag))
         {
           // Nothing to do, no tag
         }
-        else if (row.TeamCaptain.StartsWith(row.Tag, StringComparison.OrdinalIgnoreCase))
+        else if (row.TeamCaptain.StartsWith(transformedTag, StringComparison.OrdinalIgnoreCase) || row.TeamCaptain.StartsWith(row.Tag, StringComparison.OrdinalIgnoreCase))
         {
           // Nothing to do, the tag is at the default Front
         }
-        else if (row.TeamCaptain.EndsWith(row.Tag, StringComparison.OrdinalIgnoreCase))
+        else if (row.TeamCaptain.EndsWith(transformedTag, StringComparison.OrdinalIgnoreCase) || row.TeamCaptain.EndsWith(row.Tag, StringComparison.OrdinalIgnoreCase))
         {
           // Tag is at the back.
           newTeam.ClanTagOption = TagOption.Back;
         }
-        else if (row.Tag.Length >= 2)
+        else
         {
           // If the tag has 2 or more characters, check 'surrounding' criteria which is take the
           // first character of the tag and check if the captain's name begins with this character,
           // then take the last character of the tag and check if the captain's name ends with this character.
           // e.g. Tag: //, Captain's name: /captain/
-          if (row.TeamCaptain.StartsWith(row.Tag[0].ToString(), StringComparison.OrdinalIgnoreCase)
-            && row.TeamCaptain.EndsWith(row.Tag[row.Tag.Length - 1].ToString(), StringComparison.OrdinalIgnoreCase))
+          if (row.Tag.Length >= 2)
           {
-            newTeam.ClanTagOption = TagOption.Surrounding;
+            if (row.TeamCaptain.StartsWith(row.Tag[0].ToString(), StringComparison.OrdinalIgnoreCase)
+            && row.TeamCaptain.EndsWith(row.Tag[row.Tag.Length - 1].ToString(), StringComparison.OrdinalIgnoreCase))
+            {
+              newTeam.ClanTagOption = TagOption.Surrounding;
+            }
+          }
+          if (newTeam.ClanTagOption != TagOption.Surrounding && transformedTag.Length >= 2)
+          {
+            if (row.TeamCaptain.StartsWith(transformedTag[0].ToString(), StringComparison.OrdinalIgnoreCase)
+            && row.TeamCaptain.EndsWith(transformedTag[transformedTag.Length - 1].ToString(), StringComparison.OrdinalIgnoreCase))
+            {
+              newTeam.ClanTagOption = TagOption.Surrounding;
+            }
           }
         }
         // else tag is not present in the captain's name and therefore assume default front.
 
         teams.Add(newTeam);
-        CheckAndAddPlayer(row.TeamCaptain, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player2, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player3, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player4, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player5, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player6, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player7, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player8, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player9, row.Tag, newTeam, players);
-        CheckAndAddPlayer(row.Player10, row.Tag, newTeam, players);
+        CheckAndAddPlayer(row.TeamCaptain, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player2, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player3, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player4, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player5, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player6, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player7, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player8, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player9, row.Tag, transformedTag, newTeam, players);
+        CheckAndAddPlayer(row.Player10, row.Tag, transformedTag, newTeam, players);
       }
 
       return (players.ToArray(), teams.ToArray());